Update Factory.php
Raven is really cool.
It that might happen that, when an api evolves, some deprecated infos remain in its response.
What is really top on XML is the XSD, which checks the structure of the data. Json des not have something like XSD.
From openApi we can extract a structure definition. Why not using it to check an api response object's structure?
This is basically the goal of those changes, allowing the reference structure tree to be used externally.
It could be further used to check json response structure conformity. Not talking about values here, but properties being present or not, tested or ignored.